fix(tracing): overflow
